requirements.txt
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
entrypoints==0.4
executing==2.2.0
Faker==37.4.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
latest==0.6.0
lxml==6.0.0
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.13.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.46.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.5
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
